Test: Item 클레스 단위 테스트 완료했습니다.

+++ C-Shop/src/test/kotlin/com/lionTF/CShop/domain/admin/service/AdminCocktailServiceTest.kt
@@ -241,7 +241,7 @@ internal class AdminCocktailServiceTest {
         //then
         assertThat(updateCocktail.httpStatus).isEqualTo(AdminResponseDTO.toSuccessUpdateCocktailResponseDTO().httpStatus)
         assertThat(updateCocktail.message).isEqualTo(AdminResponseDTO.toSuccessUpdateCocktailResponseDTO().message)
-        assertThat(countCocktailItem).isEqualTo(itemIdList.size)
+        assertThat(countCocktailItem).isEqualTo(itemIdList.size.toLong())
     }
